feat(map): useMap hook
